closes #43. create PetT service impl

+++ pet-clinic-data/src/main/java/com/atech/repositories/PetRepository.java
@@ -3,5 +3,5 @@ package com.atech.repositories;
 import com.atech.entity.Pet;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PetRepository extends CrudRepository<Pet, Integer> {
+public interface PetRepository extends CrudRepository<Pet, Long> {
 }
