+++ pet-clinic-data/src/main/java/com/atech/repositories/SpecialityRepository.java
@@ -3,5 +3,5 @@ package com.atech.repositories;
 import com.atech.entity.Speciality;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SpecialityRepository extends CrudRepository<Speciality, Integer> {
+public interface SpecialityRepository extends CrudRepository<Speciality, Long> {
 }
